+++ CHANGELOG.md
@@ -8,9 +8,18 @@ This document follows the conventions laid out in [Keep a CHANGELOG][].
 ## Unreleased
 
 ### Added
+
+-  Support `del obj[...]` for types derived from `IList<T>` and `IDictionary<K, V>`
+
 ### Changed
 ### Fixed
 
+<<<<<<< HEAD
+=======
+-  Fixed crash when trying to `del clrObj[...]` for non-arrays
+- ci: properly exclude job (#2542)
+
+>>>>>>> 77bdf6d (implemented __delitem__ for IDictionary<K,V> and IList<T>)
 ## [3.0.5](https://github.com/pythonnet/pythonnet/releases/tag/v3.0.5) - 2024-12-13
 
 ### Added